Adds Web Servers samples
