+++ .github/CODEOWNERS
@@ -3,7 +3,6 @@ ca-certificates/ @paketo-buildpacks/utilities-maintainers
 dotnet-core/     @paketo-buildpacks/dotnet-core-maintainers
 git/             @paketo-buildpacks/utilities-maintainers
 go/              @paketo-buildpacks/go-maintainers
-httpd/           @paketo-buildpacks/web-servers-maintainers
 java/            @paketo-buildpacks/java-buildpacks
 nginx/           @paketo-buildpacks/web-servers-maintainers
 nodejs/          @paketo-buildpacks/nodejs-maintainers
@@ -11,3 +10,4 @@ php/             @paketo-buildpacks/php-maintainers
 procfile/        @paketo-buildpacks/utilities-maintainers
 python/          @paketo-buildpacks/python-maintainers
 ruby/            @paketo-buildpacks/ruby-maintainers
+web-servers/     @paketo-buildpacks/web-servers-maintainers
