--- httpd/no-config-file-sample/binding/.htpasswd
@@ -1,1 +1,0 @@
-user:$apr1$4A7us6ta$b.nzSqqXF2vkP8YbfOaH61
