+++ web-servers/javascript-frontend-sample/nginx.toml
