+++ web-servers/javascript-frontend-sample/package.json
