+++ web-servers/javascript-frontend-sample/public/favicon.ico
