+++ web-servers/nginx-sample/README.md
@@ -1,0 +1,19 @@
+# NGINX Server Sample Application
+
+## Building
+
+```bash
+pack build my-nginx-app --buildpack paketo-buildpacks/nginx
+```
+
+## Running
+
+```bash
+docker run --tty --env PORT=8080 --publish 8080:8080 my-nginx-app
+```
+
+## Viewing
+
+```bash
+curl -s localhost:8080
+```
