+++ web-servers/nginx-sample/public/index.html
