+++ web-servers/no-config-file-sample/README.md
@@ -1,0 +1,8 @@
+# Zero Configuration Static File App
+This directory hass all of the components to for a sample of a zero
+configuration static file app that is configurable with either HTTPD or NGINX
+web servers.
+
+Specific examples have been split into two seperate files.
+- For the HTTPD examples, go to the [`HTTPD.md`](HTTPD.md) file.
+- For the NGINX examples, go to the [`NGINX.md`](NGINX.md) file.
