+++ web-servers/no-config-file-sample/binding/.htpasswd
@@ -1,0 +1,1 @@
+user:$apr1$t1Apj7QI$a9M4C.h997dB2YyxgtBxA0
